# server/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/Turbines": {
      "get": {
        "tags": [
          "Turbines"
        ],
        "operationId": "Turbines_GetAll",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TurbineStatusDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/Turbines/{id}": {
      "get": {
        "tags": [
          "Turbines"
        ],
        "operationId": "Turbines_GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurbineStatusDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/Turbines/{id}/commands": {
      "get": {
        "tags": [
          "Turbines"
        ],
        "operationId": "Turbines_GetCommands",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TurbineCommand"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Turbines"
        ],
        "operationId": "Turbines_SendCommand",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendCommandRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurbineCommand"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TurbineStatusDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "turbine": {
            "$ref": "#/components/schemas/WindTurbine"
          },
          "latestMetric": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TurbineMetric"
              }
            ]
          }
        }
      },
      "WindTurbine": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "farmId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "lastSeen": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TurbineMetric": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "turbineId": {
            "type": "string"
          },
          "farmId": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "windSpeed": {
            "type": "number",
            "format": "double"
          },
          "windDirection": {
            "type": "number",
            "format": "double"
          },
          "ambientTemperature": {
            "type": "number",
            "format": "double"
          },
          "rotorSpeed": {
            "type": "number",
            "format": "double"
          },
          "powerOutput": {
            "type": "number",
            "format": "double"
          },
          "nacelleDirection": {
            "type": "number",
            "format": "double"
          },
          "bladePitch": {
            "type": "number",
            "format": "double"
          },
          "generatorTemp": {
            "type": "number",
            "format": "double"
          },
          "gearboxTemp": {
            "type": "number",
            "format": "double"
          },
          "vibration": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "TurbineCommand": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "turbineId": {
            "type": "string"
          },
          "farmId": {
            "type": "string"
          },
          "userId": {
            "type": "string",
            "format": "guid"
          },
          "userName": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "payload": {
            "type": "string",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SendCommandRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "action": {
            "type": "string"
          },
          "value": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "angle": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter your JWT token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}